+++ src/state/app_state.rs
@@ -5,6 +5,7 @@ use crate::components::stats::StatsState;
 use crate::state::boxscore::BoxscoreState;
 use crate::state::date_input::DateInput;
 use crate::state::gameday::GamedayState;
+use crate::ui::help::HelpState;
 
 /// A team must be either Home or Away.
 #[derive(Copy, Clone, Debug, Default, PartialEq)]
@@ -26,4 +27,5 @@ pub struct AppState {
     pub box_score: BoxscoreState,
     pub standings: StandingsState,
     pub stats: StatsState,
+    pub help: HelpState,
 }