+++ CHANGELOG.md
@@ -11,7 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add ABS challenge information and display in Gameday view: [PR 92](https://github.com/mlb-rs/mlbt/pull/92)
 - Add stats and standings for Spring Training: [PR 93](https://github.com/mlb-rs/mlbt/pull/93)
- 
+- Add scrolling for stats table: [PR 95](https://github.com/mlb-rs/mlbt/pull/95)
+- Add scrolling on help page: [PR 95](https://github.com/mlb-rs/mlbt/pull/95)
+  
 ### Changed
 
 - Update Rust to 1.93